decompose molgraph featurization further
